Added procedure selection
Just in case anyone wants to either preserve files, records or empty dirs.

+++ console/StorageClear.php
@@ -2,6 +2,7 @@
 
 use Storage;
 use Illuminate\Console\Command;
+use Symfony\Component\Console\Input\InputOption;
 use System\Models\File;
 
 class StorageClear extends Command
@@ -22,61 +23,68 @@ class StorageClear extends Command
      */
     public function handle()
     {
-        // remove files without related register...
-        $this->info(trans('genius.storageclear::lang.clear.seeking.files'));
+        $doAll = empty($this->option('files')) && empty($this->option('registers')) && empty($this->option('directories'));
 
-        $allFiles = Storage::allFiles('uploads');
-        $count = 0;
-        $total = count($allFiles);
-        foreach ($allFiles as $file) {
+        // remove files without related register if files option is set...
+        if($doAll || $this->option('files')) {
+            $this->info(trans('genius.storageclear::lang.clear.seeking.files'));
 
-            if (!File::where('disk_name', basename($file))->first(['id'])) {
-                Storage::delete($file);
-                $count++;
-            }
-        }
-        $this->info(trans('genius.storageclear::lang.clear.removed.files', compact('count', 'total')));
+            $allFiles = Storage::allFiles('uploads');
+            $count = 0;
+            $total = count($allFiles);
+            foreach ($allFiles as $file) {
 
-        // remove registers without file...
-        $this->info(trans('genius.storageclear::lang.clear.seeking.registers'));
-
-        $allFiles = File::all(['id', 'disk_name', 'attachment_type', 'attachment_id', 'is_public']);
-        $count = 0;
-        $total = $allFiles->count();
-
-        foreach ($allFiles as $file) {
-
-            if (!Storage::exists($file->getDiskPath())) {
-
-                $file->delete();
-                $count++;
-            } else {
-                $class = $file->attachment_type;
-                if (!$class || !class_exists($class) || !$class::find($file->attachment_id)) {
-
-                    $file->delete();
+                if (!File::where('disk_name', basename($file))->first(['id'])) {
+                    Storage::delete($file);
                     $count++;
                 }
             }
-
+            $this->info(trans('genius.storageclear::lang.clear.removed.files', compact('count', 'total')));
         }
-        $this->info(trans('genius.storageclear::lang.clear.removed.registers', compact('count', 'total')));
 
+        // remove registers without file if register option is set...
+        if($doAll || $this->option('registers')){
+            $this->info(trans('genius.storageclear::lang.clear.seeking.registers'));
 
-        // deletar pastas vazias
-        $this->info(trans('genius.storageclear::lang.clear.seeking.directories'));
+            $allFiles = File::all(['id', 'disk_name', 'attachment_type', 'attachment_id', 'is_public']);
+            $count = 0;
+            $total = $allFiles->count();
 
-        $allFolders = array_reverse(Storage::allDirectories('uploads'));
-        $count = 0;
-        $total = count($allFolders);
-        foreach ($allFolders as $directory) {
+            foreach ($allFiles as $file) {
 
-            if (!Storage::allFiles($directory)) {
-                Storage::deleteDirectory($directory);
-                $count++;
+                if (!Storage::exists($file->getDiskPath())) {
+
+                    $file->delete();
+                    $count++;
+                } else {
+                    $class = $file->attachment_type;
+                    if (!$class || !class_exists($class) || !$class::find($file->attachment_id)) {
+
+                        $file->delete();
+                        $count++;
+                    }
+                }
+
             }
+            $this->info(trans('genius.storageclear::lang.clear.removed.registers', compact('count', 'total')));
         }
-        $this->info(trans('genius.storageclear::lang.clear.removed.directories', compact('count', 'total')));
+
+        // Remove empty directories if directories option is set
+        if($doAll || $this->option('directories')) {
+            $this->info(trans('genius.storageclear::lang.clear.seeking.directories'));
+
+            $allFolders = array_reverse(Storage::allDirectories('uploads'));
+            $count = 0;
+            $total = count($allFolders);
+            foreach ($allFolders as $directory) {
+
+                if (!Storage::allFiles($directory)) {
+                    Storage::deleteDirectory($directory);
+                    $count++;
+                }
+            }
+            $this->info(trans('genius.storageclear::lang.clear.removed.directories', compact('count', 'total')));
+        }
     }
 
     /**
@@ -94,7 +102,11 @@ class StorageClear extends Command
      */
     protected function getOptions()
     {
-        return [];
+        return [
+            ['files', 'f', InputOption::VALUE_NONE, 'Remove files.', null],
+            ['registers', 'r', InputOption::VALUE_NONE, 'Remove registers.', null],
+            ['directories', 'd', InputOption::VALUE_NONE, 'Remove empty directories.', null],
+        ];
     }
 
 }
